requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
black==25.1.0
camel-ai[tools]==0.2.34
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
curl_cffi==0.10.0
distro==1.9.0
docstring-parser==0.15
eval_type_backport==0.2.2
fore==0.1.8
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy==1.15.0
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.66.5
packaging==24.2
pandas==2.2.3
pandoc==2.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
plumbum==1.9.0
ply==3.11
protobuf==6.30.1
psutil==5.9.8
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.23.1
six==1.17.0
sniffio==1.3.1
tiktoken==0.7.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
